Effective resource virtualization for dynamic IoT devices in road network:
The simulation of the proposed work performed using OMNeT++. The basic simulation setup of the proposed work virtual objects setup in Table. The simulation analysis requirements server shown the growing network size because the size of the moving node counts to gateway count and it is in the increasing order.